restore: extract batcher to another file
